+++ client/src/components/Footer.jsx
@@ -1,4 +1,3 @@
-import { Link } from "react-router-dom";
 
 export default function Footer() {
   return (
